[#33] FilterAttributeType 내 image 프로퍼티 추가
